requirements.txt
Flask>=3.0.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.1
smolagents>=1.0.0
litellm>=1.50.0
gunicorn
azure-storage-blob>=12.19.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pandas>=2.0.0

# Configuration and validation
pydantic>=2.0.0
pyyaml>=6.0

# Development tools
black>=24.0.0
flake8>=7.0.0
tensorboard>=2.15.0
mlflow>=2.9.0
